Code improvements
Using booleans in config.lua (strings are still supported for backwards
compatibility)
Improved maintainability of existing enums
Removed unused enums
Reduced memory footprint of item abilities

+++ src/tools.h
@@ -79,8 +79,8 @@ bool booleanString(const std::string& str);
 
 std::string getWeaponName(WeaponType_t weaponType);
 
-uint32_t combatTypeToIndex(CombatType_t combatType);
-CombatType_t indexToCombatType(uint32_t v);
+size_t combatTypeToIndex(CombatType_t combatType);
+CombatType_t indexToCombatType(size_t v);
 
 uint8_t serverFluidToClient(uint8_t serverFluid);
 uint8_t clientFluidToServer(uint8_t clientFluid);
